opentelemetrybased/python/requirements.txt
click==8.0.1
Deprecated==1.2.13
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
opentelemetry-api==1.5.0
opentelemetry-instrumentation==0.24b0
opentelemetry-sdk==1.5.0
opentelemetry-semantic-conventions==0.24b0
prometheus-client==0.11.0
Werkzeug==2.0.1
wrapt==1.12.1
